feat(web): lazy load Recharts in Analytics Dashboard to reduce bundle size
Closes #1426
